* commit '6cd4daa63dd7206efc75391efee01658bb8cec2e':
  删除max、bpcatelogpage、pharbers-web等项目
  删除多余项目
  启动资源 notice修改
  删除项目完成
  修改
  临时提交
  临时提交
  创建project
  sample调试
  临时提交创建问题
  修改s3上传 iot
  sample 数据采样修改

+++ client-helper/project-model-helper/src/eventHandler/phAnalyzeDataListHomeEventHandler.js
@@ -3,6 +3,7 @@ import { hostName } from "../config/envConfig"
 // eslint-disable-next-line no-unused-vars
 export async function phAnalyzeDataListHomeEventHandler(e, route) {
 	let params = e.detail[0].args.param
+	const createResourceEventName = "resource_create"
 	const startUrl = `${hostName}/phresourceaction`
 	const accessToken = route.cookies.read("access_token")
 	let uri = "/projects"
@@ -73,21 +74,38 @@ export async function phAnalyzeDataListHomeEventHandler(e, route) {
 			break
 		case "deleteProject":
 			if (params) {
-				let uri = `${hostName}/phcreateproject/projects/${params.projectId}`
-				let results = await fetch(uri, {
-					method: "delete",
-					headers: {
-						Authorization: accessToken,
-						"Content-Type": "application/vnd.api+json",
-						Accept: "application/vnd.api+json"
+				/**
+				 * 1.判断是否正在启动
+				 * 2.删除resource
+				 * 3.删除project
+				 */
+				let bool = await checkStartResourceFun("del")
+				if (bool) {
+					let delResourceUri = `${hostName}/phcreateproject/resources/${params.resourceId}`
+					await fetch(delResourceUri, {
+						method: "delete",
+						headers: {
+							Authorization: accessToken,
+							"Content-Type": "application/vnd.api+json",
+							Accept: "application/vnd.api+json"
+						}
+					})
+					let uri = `${hostName}/phcreateproject/projects/${params.projectId}`
+					let results = await fetch(uri, {
+						method: "delete",
+						headers: {
+							Authorization: accessToken,
+							"Content-Type": "application/vnd.api+json",
+							Accept: "application/vnd.api+json"
+						}
+					})
+					if (results.status === 204) {
+						alert("删除项目成功！")
+						window.location.href =
+							"https://general.pharbers.com/projects"
+					} else {
+						alert("删除失败！")
 					}
-				})
-				if (results.status === 204) {
-					alert("删除项目成功！")
-					window.location.href =
-						"https://general.pharbers.com/projects"
-				} else {
-					alert("删除失败！")
 				}
 			}
 			break
@@ -109,11 +127,21 @@ export async function phAnalyzeDataListHomeEventHandler(e, route) {
 				let checked = await checkStartResourceFun()
 				if (checked) {
 					console.log("开始启动")
+					// route.noticeService.defineAction({
+					// 	type: "iot",
+					// 	id: "resource_create",
+					// 	projectId: params.projectId,
+					// 	ownerId: "*",
+					// 	callBack: callback
+					// })
 					route.noticeService.defineAction({
 						type: "iot",
-						id: "resource_create",
+						remoteResource: "notification",
+						runnerId: "",
+						id: result.data.id,
+						eventName: createResourceEventName,
 						projectId: params.projectId,
-						ownerId: "*",
+						ownerId: route.cookies.read("account_id"),
 						callBack: callback
 					})
 					let body = {
@@ -158,7 +186,7 @@ export async function phAnalyzeDataListHomeEventHandler(e, route) {
 		route.loadingService.loading.style.display = "none"
 	}
 
-	async function checkStartResourceFun() {
+	async function checkStartResourceFun(startMsg) {
 		const startUrl = `${hostName}/phresourceaction`
 		let startBody = {
 			projectName: params.projectName,
@@ -183,15 +211,28 @@ export async function phAnalyzeDataListHomeEventHandler(e, route) {
 		)
 		console.log("判断是否启动：", startResults)
 		if (startResults.data.resource_status === "starting") {
+			// route.noticeService.defineAction({
+			// 	type: "iot",
+			// 	id: "resource_create",
+			// 	projectId: params.projectId,
+			// 	ownerId: "*",
+			// 	callBack: callback
+			// })
 			route.noticeService.defineAction({
 				type: "iot",
-				id: "resource_create",
+				remoteResource: "notification",
+				runnerId: "",
+				id: result.data.id,
+				eventName: createResourceEventName,
 				projectId: params.projectId,
-				ownerId: "*",
+				ownerId: route.cookies.read("account_id"),
 				callBack: callback
 			})
 			route.resourceActionService.boolChecked = false
 			route.loadingService.loading.style.display = "flex"
+			if (startMsg === "del") {
+				alert("正在启动，不能删除项目")
+			}
 			return false
 		} else if (startResults.data.resource_status === "started") {
 			e.detail[0].args.element.showStartButton = false //按钮disabled
